chore(worker/staking): move tag from transaction to exchange
