+++ service/indexer/internal/worker/exchange/staking/platform.go
