+++ service/indexer/internal/worker/exchange/staking/worker_rss3.go
@@ -111,8 +111,8 @@ func (s *Staking) buildRSS3StakingTransfer(ctx context.Context, _transaction mod
 		TransactionHash: _transaction.Hash,
 		Timestamp:       _transaction.Timestamp,
 		BlockNumber:     big.NewInt(_transaction.BlockNumber),
-		Tag:             filter.TagTransaction,
-		Type:            filter.TransactionStaking,
+		Tag:             filter.ExchangeSwap,
+		Type:            filter.ExchangeStaking,
 		Index:           int64(log.Index),
 		AddressFrom:     _transaction.AddressFrom,
 		AddressTo:       _transaction.AddressTo,
